Example of another tool that generates the schema
This is a pretty new  tool, that I've found out a coupld of weeks ago,
and it's doing what I wanted to implement for values schema. It's
reading special comments (annotations) in the helm values file and then
it generates a schema.

Why do I find it better than the helm-schema plugin, that we use
currently?

We have generated the current schema with the plugin, but since the
`servce.nodePort` is set to null in the values, the type of this field
is null, hence, it's not possible to use any value.

With the helm-schema (tool), it's possible to add an annotation in the
values, so it know that it can either be an `integer` of `null`. Please,
check the changes.